yhdg-front-server/src/test/java/cn/com/yusong/yhdg/frontserver/service/yms/TerminalUploadLogServiceTest.java
package cn.com.yusong.yhdg.frontserver.service.yms;

import cn.com.yusong.yhdg.common.domain.basic.Agent;
import cn.com.yusong.yhdg.common.domain.basic.Partner;
import cn.com.yusong.yhdg.common.domain.yms.Playlist;
import cn.com.yusong.yhdg.common.domain.yms.Terminal;
import cn.com.yusong.yhdg.common.domain.yms.TerminalStrategy;
import cn.com.yusong.yhdg.common.domain.yms.TerminalUploadLog;
import cn.com.yusong.yhdg.frontserver.BaseJunit4Test;
import org.junit.Test;
import org.springframework.beans.factory.annotation.Autowired;

import static org.junit.Assert.*;

public class TerminalUploadLogServiceTest extends BaseJunit4Test {
    @Autowired
    TerminalUploadLogService terminalUploadLogService;

    @Test
    public void updateStatus() {
        Partner partner = newPartner();
        insertPartner(partner);

        Agent agent = newAgent(partner.getId());
        insertAgent(agent);

        Playlist playlist = newPlaylist(agent.getId());
        insertPlaylist(playlist);

        TerminalStrategy terminalStrategy = newTerminalStrategy(agent.getId());
        insertTerminalStrategy(terminalStrategy);

        Terminal terminal = newTerminal(agent.getId(), terminalStrategy.getId(), playlist.getId());
        insertTerminal(terminal);

        TerminalUploadLog terminalUploadLog = newTerminalUploadLog(agent.getId(), terminal.getId());
        insertTerminalUploadLog(terminalUploadLog);

        assertEquals(1, terminalUploadLogService.updateStatus(terminalUploadLog));
    }

    @Test
    public void update() {
        Partner partner = newPartner();
        insertPartner(partner);

        Agent agent = newAgent(partner.getId());
        insertAgent(agent);

        Playlist playlist = newPlaylist(agent.getId());
        insertPlaylist(playlist);

        TerminalStrategy terminalStrategy = newTerminalStrategy(agent.getId());
        insertTerminalStrategy(terminalStrategy);

        Terminal terminal = newTerminal(agent.getId(), terminalStrategy.getId(), playlist.getId());
        insertTerminal(terminal);

        TerminalUploadLog terminalUploadLog = newTerminalUploadLog(agent.getId(), terminal.getId());
        insertTerminalUploadLog(terminalUploadLog);

        assertEquals(1, terminalUploadLogService.update(terminalUploadLog));
    }
}